requirements.txt
azure-search-documents==11.6.0b4
azure-storage-blob==12.20.0
azure-identity==1.16.0
aiohttp
ipykernel
python-dotenv
requests
openai
streamlit==1.38.0
httpx==0.27.2
pymupdf